// src/enums/actionTypes.enum.ts
export enum ActionTypesEnum {
  START_GAME = "START_GAME",
  PLACE_SHIP = "PLACE_SHIP",
  SET_MY_BOARD = "SET_MY_BOARD",
  SET_PLAYER_NAME = "SET_PLAYER_NAME",
  SET_OPPONENT_NAME = "SET_OPPONENT_NAME",
  SET_OPPONENT_BOARD = "SET_OPPONENT_BOARD",
  SET_LOBBY_CODE = "SET_LOBBY_CODE",
  HIT = "HIT",
  MISS = "MISS",
  SET_READY = "SET_READY",
  SET_CURRENT_PLAYER = "SET_CURRENT_PLAYER",
  SET_OPPONENT_READY = "SET_OPPONENT_READY",
  SET_WINNER = "SET_WINNER",
  RESET_GAME = "RESET_GAME",
}